More astheticly pleasing "Incorrect" message.
Instead of just printing incorrect and re-asking the question, used the built-in prompt parameter key "validate" to validate the question. No need to print any checkmark or cross emoji now. correct_on_first_try also is now a property of the class so it can be modified on a built-in class "validate" function.